Pages added for Native View Access Request and Emails
