+++ src/test/java/appsCommon/PageInitializer.java
@@ -76,12 +76,7 @@ import ServiceNow.NERD.Pages.NERDKnowledgebasePage;
 import ServiceNow.NERD.Pages.NERDSubmissionsPage;
 import ServiceNow.NERD.StepsImplementation.NERDLoginStepsImplementation;
 import ServiceNow.NERD.StepsImplementation.NativeViewImpersonateUser;
-import ServiceNow.SEER.Pages.NativeViewSentViewPage;
-import ServiceNow.SEER.Pages.SEERDataAccessRequestPage;
-import ServiceNow.SEER.Pages.SEERExistingAccountPage;
-import ServiceNow.SEER.Pages.SEERIncidenceDatabaseDetailsPage;
-import ServiceNow.SEER.Pages.SEERLandingPage;
-import ServiceNow.SEER.Pages.SEERUserRegistrationPage;
+import ServiceNow.SEER.Pages.*;
 import ServiceNow.SEER.StepsImplementation.SEERDataAccessRequestPageStepsImpl;
 import CustomBusiness.DCEG.Pages.AdminFlowPage;
 import CustomBusiness.DCEG.Pages.BranchAdminPage;
@@ -204,6 +199,8 @@ public class PageInitializer {
 	protected static SEERUserRegistrationPage seerUserRegistrationPage;
 	protected static SEERDataAccessRequestPage seerDataAccessRequestPage;
 	protected static NativeViewSentViewPage nativeViewSentViewPage;
+	protected static NativeViewAccessRequestPage nativeViewAccessRequestPage;
+	protected static NativeViewEmailsPage nativeViewEmailsPage;
 	protected static SEERIncidenceDatabaseDetailsPage seerIncidenceDatabaseDetailsPage;
 	protected static SEERExistingAccountPage seerExistingAccountPage;
 	protected static SEERDataAccessRequestPageStepsImpl seerDataAccessRequestPageStepsImpl;
@@ -322,6 +319,8 @@ public class PageInitializer {
 		seerUserRegistrationPage = new SEERUserRegistrationPage();
 		seerDataAccessRequestPage = new SEERDataAccessRequestPage();
 		nativeViewSentViewPage = new NativeViewSentViewPage();
+		nativeViewAccessRequestPage = new NativeViewAccessRequestPage();
+		nativeViewEmailsPage = new NativeViewEmailsPage();
 		seerIncidenceDatabaseDetailsPage = new SEERIncidenceDatabaseDetailsPage();
 		nerdLoginStepsImplementation = new NERDLoginStepsImplementation();
 		nerdCrsKnowledgeDatabaseSubmissionsPage = new NERDSubmissionsPage();
